ENH: Add datalad installation via conda
It will also install git-annex which datalad uses for management of large
files.

Since it is just a "tool", I have not used --freeze-installed to benefit from
possible upgrades etc.

It could also be installed from stock ubuntu, but would be an older version.
Also backports are available from NeuroDebian, but I have decided to just go with
conda installation for now.

Closes #50